+++ content/webapp/components/ItemRequestModal/ItemRequestModal.CurrentRequests.tsx
@@ -3,17 +3,13 @@ import styled from 'styled-components';
 
 import Space from '@weco/common/views/components/styled/Space';
 
-export const CTAs = styled(Space).attrs({
-  $v: { size: 'l', properties: ['margin-top'] },
-})``;
-
 const CurrentRequestCount = styled(Space).attrs({
   $h: { size: 's', properties: ['padding-left', 'margin-left'] },
 })`
   border-left: 5px solid ${props => props.theme.color('yellow')};
 `;
 
-export const CurrentRequests: FunctionComponent<{
+const CurrentRequests: FunctionComponent<{
   allowedHoldRequests: number;
   currentHoldRequests?: number;
 }> = ({ allowedHoldRequests, currentHoldRequests }) =>
@@ -27,10 +23,4 @@ export const CurrentRequests: FunctionComponent<{
     </CurrentRequestCount>
   ) : null;
 
-export const Header = styled(Space).attrs({
-  $v: { size: 'm', properties: ['margin-bottom'] },
-})`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
+export default CurrentRequests;
