+++ content/webapp/components/ItemRequestModal/RequestDialog/index.tsx
@@ -8,14 +8,21 @@ import { allowedRequests } from '@weco/common/values/requests';
 import Button, { ButtonTypes } from '@weco/common/views/components/Buttons';
 import Space from '@weco/common/views/components/styled/Space';
 import { themeValues } from '@weco/common/views/themes/config';
-import RequestingDayPicker from '@weco/content/components/RequestingDayPicker';
+import CurrentRequests from '@weco/content/components/ItemRequestModal/ItemRequestModal.CurrentRequests';
+import {
+  dateAsValue,
+  dateFromValue,
+} from '@weco/content/components/ItemRequestModal/ItemRequestModal.helpers';
+import {
+  CTAs,
+  Header,
+} from '@weco/content/components/ItemRequestModal/ItemRequestModal.styles';
 import {
   PhysicalItem,
   Work,
 } from '@weco/content/services/wellcome/catalogue/types';
 
-import { CTAs, CurrentRequests, Header } from './common';
-import { dateAsValue, dateFromValue } from './format-date';
+import RequestingDayPicker from './RequestDialog.RequestingDayPicker';
 
 const PickUpDate = styled(Space).attrs({
   $v: { size: 'l', properties: ['padding-top', 'padding-bottom'] },
